fix: 중복되는 데이터 삭제
