+++ frontend/src/constants/tabs.ts
@@ -1,0 +1,33 @@
+import { Tab } from '@/components/common/Tabs/Tabs';
+
+export const newChecklistTabs: Tab[] = [
+  { id: 0, name: '기본 정보' },
+  {
+    id: 1,
+    name: '청결',
+  },
+  {
+    id: 2,
+    name: '방 컨디션',
+  },
+  {
+    id: 3,
+    name: '편의시설',
+  },
+  {
+    id: 4,
+    name: '옵션',
+  },
+  {
+    id: 5,
+    name: '주거환경',
+  },
+  {
+    id: 6,
+    name: '보안',
+  },
+  {
+    id: 7,
+    name: '경제적',
+  },
+];
