+++ frontend/src/pages/NewChecklistPage/NewChecklistPage.tsx
@@ -6,9 +6,9 @@ import { postChecklist } from '@/apis/checklist';
 import Button from '@/components/common/Button/Button';
 import Header from '@/components/common/Header/Header';
 import { TabProvider } from '@/components/common/Tabs/TabContext';
-import { Tab } from '@/components/common/Tabs/Tabs';
 import { useToastContext } from '@/components/common/Toast/ToastContext';
 import { ROUTE_PATH } from '@/constants/routePath';
+import { newChecklistTabs } from '@/constants/tabs';
 import useChecklistAnswer from '@/hooks/useChecklistAnswer';
 import useInputs from '@/hooks/useInput';
 import NewChecklistBody from '@/pages/NewChecklistPage/NewChecklistBody';
@@ -29,38 +29,6 @@ const DefaultRoomInfo: RoomInfo = {
   realEstate: '방끗공인중개사',
 };
 
-export const newChecklistTabs: Tab[] = [
-  { id: 0, name: '기본 정보' },
-  {
-    id: 1,
-    name: '청결',
-  },
-  {
-    id: 2,
-    name: '방 컨디션',
-  },
-  {
-    id: 3,
-    name: '편의시설',
-  },
-  {
-    id: 4,
-    name: '옵션',
-  },
-  {
-    id: 5,
-    name: '주거환경',
-  },
-  {
-    id: 6,
-    name: '보안',
-  },
-  {
-    id: 7,
-    name: '경제적',
-  },
-];
-
 const NewChecklistPage = () => {
   const { showToast } = useToastContext();
 
@@ -71,7 +39,6 @@ const NewChecklistPage = () => {
   const [selectedOptions, setSelectedOptions] = useState<number[]>([]);
 
   /*체크리스트 답변*/
-
   const { addAnswer, deleteAnswer, questionSelectedAnswer, checklistAnswers } = useChecklistAnswer();
 
   const navigate = useNavigate();
